backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
pymongo==4.6.0
cassandra-driver==3.29.1
redis==5.0.1
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
scipy==1.11.4
